requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# File Upload & Processing
python-multipart==0.0.6
Pillow==10.2.0
pillow-heif==0.18.0

# OCR Engines
google-cloud-vision==3.5.0
pytesseract==0.3.10

# Rate Limiting & Caching
slowapi==0.1.9
cachetools==5.3.2

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
